python/web3-cli/requirements.txt
web3==6.15.1
eth-account==0.11.0
py-solc-x==2.0.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pytest==7.4.3
